update tests to work correctly, remove front end tests temporarily(already broken)
