+++ package.json
@@ -7,7 +7,7 @@
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
     "start": "node server/index.js",
-    "test": "jest",
+    "test": "jest --detectOpenHandles",
     "testQuietly": "jest --silent",
     "lint": "eslint client/src/*.jsx database/db.js server/index.js"
   },
